Admin JobRole Qualification added,New Structure

+++ app/Jobs/RemoveJobRole.php
@@ -38,15 +38,12 @@ class RemoveJobRole implements ShouldQueue
         
         $jobrole = JobRole::find($request->id);
         if ($jobrole) {
-
-            // if (count($jobrole->job_roles) == 0) {
+            if (count($jobrole->partners) > 0) {
+                return response()->json(array('type' => 'error', 'message' => "You cannot <span style='font-weight:bold;color:red'>Remove</span> a Job Role while its associated with some Training Partners"),200);
+            } else {
                 $jobrole->delete();
-                return response()->json(array('type' => 'success', 'message' => "Expository <span style='font-weight:bold;color:blue'>$request->name</span> has been Removed Successfully"),200);
-            // } else {
-            //     return response()->json(array('type' => 'error', 'message' => "You cannot <span style='font-weight:bold;color:red'>Remove</span> any Expository while its associated with some Job Roles"),200);
-            // }
-            
-        
+                return response()->json(array('type' => 'success', 'message' => "Job Role <span style='font-weight:bold;color:blue'>$request->name</span> has been Removed Successfully"),200);
+            }
         } else {
             return response()->json(array('type' => 'error', 'message' => "<span style='font-weight:bold;color:red'>$request->name</span> Expository is not Found"),200);
         }
